+++ python/dune/__main__.py
@@ -1,6 +1,6 @@
 import sys
 from argparse import ArgumentParser
-from dune.commands import configure, checkbuilddirs, rmgenerated, listdunetype, listgenerated
+from dune.commands import printinfo, configure, listgenerated, rmgenerated, fixdunepy, listdunetype, checkbuilddirs
 
 # NOTE: do not import from dune.common (and consequently from dune.generator)
 #       at top level to avoid failure due to missing mpi4py.
@@ -10,6 +10,10 @@ def run(arguments=None):
     parser = ArgumentParser(description='Execute DUNE commands', prog='dune')
     subparsers = parser.add_subparsers(dest='command')
 
+    # Info
+    parserConfigure = subparsers.add_parser('info',
+              help='Print information about dune-py')
+
     # Configure
     parserConfigure = subparsers.add_parser('configure',
               help='Tag dune-py to be reconfigured before next use')
@@ -18,6 +22,8 @@ def run(arguments=None):
     parserList = subparsers.add_parser('list', help='List all generated modules')
     parserList.add_argument('--alphabetical', dest='sort', action='store_const', const='alphabetical', default='bydate',
               help='List modules in alphabetical order (default: by date)')
+    parserList.add_argument('--ccfiles', dest='ccfiles', action='store_const', const=True, default=False,
+              help='List modules by .cc ending instead of .so')
 
     # Remove
     parserRemove = subparsers.add_parser('remove', help='Remove generated modules')
@@ -25,6 +31,12 @@ def run(arguments=None):
               help='Instead of a pattern provide a date to remove all modules not having been loaded after that date')
     parserRemove.add_argument('modules', nargs='*',  default=[],
               help='Patterns of modules ("*.cc" and dune-py path is added to each argument) or "all"')
+
+    # Fix dune-py
+    parserFix = subparsers.add_parser('fix-dunepy',
+              help='Find inconsistencies in dune-py and try to fix automatically. This will potentially delete all generated modules.')
+    parserFix.add_argument('--force', dest='force', action='store_const', const=True, default=False,
+              help='force complete removal of dune-py without checking for inconsistencies')
 
     # Dunetype
     parserDunetype = subparsers.add_parser('dunetype', help='Show dune types for given modules')
@@ -39,11 +51,15 @@ def run(arguments=None):
 
     ret = 0
     args = parser.parse_args(arguments)
-    if args.command == 'configure':
+
+    if args.command == 'info':
+        ret = printinfo()
+
+    elif args.command == 'configure':
         ret = configure()
 
-    elif args.command == 'checkbuilddirs':
-        ret = checkbuilddirs(args.args)
+    elif args.command == 'list':
+        ret = listgenerated(args.sort, args.ccfiles)
 
     elif args.command == 'remove':
         if args.modules == []:
@@ -51,14 +67,17 @@ def run(arguments=None):
         else:
             ret = rmgenerated(args.modules, args.date)
 
+    elif args.command == 'fix-dunepy':
+        ret = fixdunepy(args.force)
+
     elif args.command == 'dunetype':
         if args.modules == []:
             parserDunetype.print_help()
         else:
             ret = listdunetype(args.modules)
 
-    elif args.command == 'list':
-        ret = listgenerated(args.sort)
+    elif args.command == 'checkbuilddirs':
+        ret = checkbuilddirs(args.args)
 
     else:
         parser.print_help()