+++ python/dune/commands.py
@@ -1,8 +1,19 @@
-import os, glob, time
+import os, glob, time, fileinput, subprocess, shutil
 from dune.packagemetadata import getBuildMetaData, forceConfigure
 
 # NOTE: do not import from dune.common (and consequently from dune.generator)
 #       at top level to avoid failure due to missing mpi4py.
+
+def printinfo():
+    from dune.common.module import getDunePyDir
+    dune_py_dir = getDunePyDir()
+    print("Location of dune-py:", dune_py_dir)
+
+    generated_dir = os.path.join(dune_py_dir, 'python', 'dune', 'generated')
+    files = glob.glob(os.path.join(generated_dir, '*.so'))
+    print(len(files), "generated modules")
+    return 0
+
 
 def configure():
     print('Set up dune-py module for reconfiguration')
@@ -41,15 +52,70 @@ def rmgenerated(args, date):
     return 0
 
 
-def listgenerated(args):
+def fixdunepy(force):
+    from dune.common.module import getDunePyDir
+    from dune.generator.remove import removeGenerated
+    dune_py_dir = getDunePyDir()
+    generated_dir = os.path.join(dune_py_dir, 'python', 'dune', 'generated')
+
+    def filebases(list):
+        return [os.path.splitext(os.path.basename(l))[0] for l in list]
+
+    if force:
+        if os.path.isdir(dune_py_dir):
+            shutil.rmtree(dune_py_dir)
+    else:
+        ccfiles = filebases( glob.glob(os.path.join(generated_dir, '*.cc')) )
+        sofiles = filebases( glob.glob(os.path.join(generated_dir, '*.so')) )
+        cmakedirs = filebases( glob.glob(os.path.join(generated_dir, "CMakeFiles", "*.dir") ) )
+
+        # Search CMakeLists.txt for pattern: dune_add_pybind11_module(NAME module_123 EXCLUDE_FROM_ALL)
+        bracket = ['dune_add_pybind11_module(NAME ', ' EXCLUDE_FROM_ALL']
+        cmakeentries = []
+        for line in fileinput.input( os.path.join(generated_dir, 'CMakeLists.txt') ):
+            start = line.find(bracket[0])
+            if start != -1:
+              start = start + len(bracket[0])
+              end = line.find(bracket[1], start)
+              cmakeentries += [line[start:end]]
+
+        occurences = {}
+        for i in ccfiles + sofiles + cmakedirs + cmakeentries:
+            occurences[i] = occurences.get(i, 0) + 1
+
+        inconsistent = []
+        for o in occurences:
+            if occurences[o] != 4:
+                inconsistent += [o]
+
+        if len(inconsistent) > 0:
+            print("Fix inconsistencies:")
+            for i in inconsistent:
+                print(" ", i)
+            removeGenerated(inconsistent, verbose=False)
+
+        # Call 'cmake .' and check output
+        if os.path.isdir(dune_py_dir):
+            p = subprocess.run(["cmake", "."], cwd=dune_py_dir, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+            if p.returncode != 0:
+                print("CMake failed! Remove whole dune-py.")
+                shutil.rmtree(dune_py_dir)
+
+    return 0
+
+def listgenerated(sort, ccfiles):
     from dune.common.module import getDunePyDir
     dune_py_dir = getDunePyDir()
     generated_dir = os.path.join(dune_py_dir, 'python', 'dune', 'generated')
 
-    files = glob.glob(os.path.join(generated_dir, '*.so'))
-    if args == 'bydate':
+    if ccfiles:
+        files = glob.glob(os.path.join(generated_dir, '*.cc'))
+    else:
+        files = glob.glob(os.path.join(generated_dir, '*.so'))
+
+    if sort == 'bydate':
         files.sort(key=os.path.getatime)
-    elif args == 'alphabetical':
+    elif sort == 'alphabetical':
         files.sort()
 
     for filename in files: